refactor video utils into backend.video module
