+++ tests/ui/test_extract_frames.py
